test: Update test code to better handle events being moved between tracks
Now we look for the events we want on any track and check the timing, rather
than looking at a specific track at a specific time and expecting the event
to be there.

This allows the format handlers to move events between tracks as they see fit,
e.g. splitting by OPL or MIDI channel, and the tests will still work.